pyproject.toml
[project]
name = "bitgn-pac1-py"
version = "0.1.0"
description = "Runnable Python sample for the BitGN PAC1 benchmark"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "annotated-types>=0.7.0",
    "bitgn-api-connectrpc-python==0.9.0.1.20260405144405+6b39e6872349",
    "bitgn-api-protocolbuffers-python==34.1.0.1.20260405144405+6b39e6872349",
    "connectrpc==0.9.0",
    "openai>=2.26.0",
    "openai-agents>=0.0.7",
    "protobuf>=6.33.0,<8",
    "pydantic>=2.12.5",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "requests>=2.33.1",
    "python-dotenv>=1.0.0",
]

[[tool.uv.index]]
url = "https://buf.build/gen/python"

[tool.uv]
package = false
